aider: Add db folder to gitignore.
# Aider chat conversation:

USER: also add the db folder to gitignore

ASSISTANT: Sure, we can add the db folder to the .gitignore file to exclude it from being committed to the repository.

Here is the proposed change:

```gitignore
.gitignore
<<<<<<< ORIGINAL
.aider*
.env
=======
.aider*
.env
/db/
>>>>>>> UPDATED
```

This change will add a new line to the .gitignore file to exclude the db folder.